Update PROBLEM-TRACKS
Primality-Test
A prime number is a number which is only divisible by 1 and itself.
Given number N check if it is prime or not.

YOUR TASK:
You don't need to read input or print anything. Your task is to complete the function isPrime() that takes N as input parameter and returns True if N is prime else returns False. 

Expected Time Complexity : O(N1/2)
Expected Auxilliary Space :  O(1)

CONSTRAINTS:
1 <= N <= 109